+++ ActionTowerDefense/src/ScreenTextUI.cpp
@@ -4,7 +4,7 @@
 #include "GDIRenderer.h"
 #include "RenderManager.h"
 
-ScreenTextUI::ScreenTextUI(const std::wstring& text, const Vector2& position, const Gdiplus::Color& color, float size)
+ScreenTextUI::ScreenTextUI(const std::wstring& text, const Vector2& position, const Gdiplus::Color& color, int size)
 	: m_Text(text), m_Color(color), m_Size(size)
 {
 	m_Position = position;
